refactor: JobRegisterRadioGroup에 getRadioProps 적용
