+++ front-end/src/domains/JobRegisterForm/JobRegisterForm.tsx
@@ -1,9 +1,10 @@
 import { FormHTMLAttributes, useState } from "react";
-import { usePostJobRegister } from "../../hooks";
-import useJobRegisterForm from "./useJobRegisterForm";
+import { expectedTimeValidator, jobNameValidator, minPerformanceValidator } from "./validator";
+import { getFormProps, getInputProps, useForm, usePostJobRegister } from "../../hooks";
 import { Alert, Button, Dimmer, Input, Loading, Text } from "../../components";
 import { StyledForm } from "./JobRegisterForm.styled";
 import JobRegisterRadioGroup from "../JobRegisterRadioGroup/JobRegisterRadioGroup";
+import { Values } from "./JobRegisterForm.type";
 
 interface JobRegisterFormProps extends FormHTMLAttributes<HTMLFormElement> {
   labId: number;
@@ -13,15 +14,51 @@ const JobRegisterForm = ({ labId, ...rest }: JobRegisterFormProps) => {
   const { status, makeRequest, done } = usePostJobRegister({ labId });
   const [key, setKey] = useState(0);
 
-  const {
-    form,
-    reset,
-    jobNameInputProps,
-    expectedTimeInputProps,
-    minPerformanceInputProps,
-    metaDataInputProps,
-    gpuServerSelectProps,
-  } = useJobRegisterForm(makeRequest);
+  const { state, dispatch, reset } = useForm<Values>({
+    jobName: "",
+    expectedTime: "",
+    minPerformance: "" as unknown as number,
+    metaData: "",
+    gpuServerId: "" as unknown as number,
+  });
+
+  const handleSubmit = ({ jobName, expectedTime, gpuServerId, metaData }: Values) => {
+    // eslint-disable-next-line @typescript-eslint/no-floating-promises
+    makeRequest({ name: jobName, expectedTime, gpuServerId, metaData });
+  };
+
+  const form = getFormProps({ state, dispatch, handleSubmit });
+
+  const jobNameInputProps = getInputProps({
+    state,
+    dispatch,
+    name: "jobName",
+    label: "Job 이름",
+    validator: jobNameValidator,
+  });
+
+  const expectedTimeInputProps = getInputProps({
+    state,
+    dispatch,
+    name: "expectedTime",
+    label: "Job 예상 실행 시간(hour)",
+    validator: expectedTimeValidator,
+  });
+
+  const minPerformanceInputProps = getInputProps({
+    state,
+    dispatch,
+    name: "minPerformance",
+    label: "최소 필요 연산량(TFLOPS)",
+    validator: minPerformanceValidator,
+  });
+
+  const metaDataInputProps = getInputProps({
+    state,
+    dispatch,
+    name: "metaData",
+    label: "Docker Hub Url",
+  });
 
   const handleConfirm = () => {
     reset();
@@ -59,7 +96,11 @@ const JobRegisterForm = ({ labId, ...rest }: JobRegisterFormProps) => {
         <Input size="sm" {...minPerformanceInputProps} />
         <Input size="sm" {...metaDataInputProps} />
         <JobRegisterRadioGroup
-          {...gpuServerSelectProps}
+          labId={labId}
+          state={state}
+          dispatch={dispatch}
+          name="gpuServerId"
+          label="GPU 서버 선택"
           minPerformance={Number(minPerformanceInputProps.value)}
         />
         <Button className="submit" color="secondary" disabled={status !== "idle"}>
