+++ front-end/src/domains/JobRegisterForm/JobRegisterForm.type.ts
@@ -1,0 +1,7 @@
+export type Values = {
+  jobName: string;
+  expectedTime: string;
+  minPerformance: number;
+  metaData: string;
+  gpuServerId: number;
+};
